refactor(Contact): support nullable address
