+++ src/Features/Home/components/Contact/Contact.tsx
@@ -1,5 +1,6 @@
-import useContact from '../../hooks/useContact';
-import SocialMedia from './SocialMedia';
+import useContact from '../../../hooks/useContact';
+import SocialMedia from '../SocialMedia';
+import ContactInfo from './components/ContactInfo';
 
 // const initialState = {
 //   name: '',
@@ -96,33 +97,9 @@ export default function Contact() {
               </form>
             </div>
           </div>
-          <div className="col-md-3 col-md-offset-1 contact-info">
-            <div className="contact-item">
-              <h3>Contact Info</h3>
-              <p>
-                <span>
-                  <i className="fa fa-map-marker"></i> Address
-                </span>
-                {contact?.address}
-              </p>
-            </div>
-            <div className="contact-item">
-              <p>
-                <span>
-                  <i className="fa fa-phone"></i> Phone
-                </span>{' '}
-                {contact?.phone}
-              </p>
-            </div>
-            <div className="contact-item">
-              <p>
-                <span>
-                  <i className="fa fa-envelope-o"></i> Email
-                </span>{' '}
-                {contact?.email}
-              </p>
-            </div>
-          </div>
+
+          <ContactInfo />
+
           <div className="col-md-12">
             <div className="row">
               <SocialMedia />
