+++ src/Features/Home/components/Contact/index.tsx
@@ -1,0 +1,1 @@
+export { default } from './Contact';
